+++ nequip/nn/embedding/__init__.py
@@ -1,3 +1,4 @@
+from .node import NodeTypeEmbed
 from ._one_hot import OneHotAtomEncoding
 from ._edge import (
     EdgeLengthNormalizer,
@@ -8,6 +9,7 @@ from ._edge import (
 from .cutoffs import PolynomialCutoff
 
 __all__ = [
+    NodeTypeEmbed,
     OneHotAtomEncoding,
     EdgeLengthNormalizer,
     BesselEdgeLengthEncoding,